refactor: separate Result to ResultRepos and ResultUser
