--- src/components/Result/index.ts
@@ -1,3 +1,0 @@
-import ResultRepos from './Result';
-
-export default ResultRepos;
