+++ src/components/ResultRepos/index.ts
@@ -1,0 +1,3 @@
+import ResultRepos from './ResultRepos';
+
+export default ResultRepos;
