+++ src/components/ResultUser/ResultUser.module.scss
@@ -1,0 +1,99 @@
+.result {
+  &__title {
+    margin: 1.1rem 0;
+    font-size: 1.5rem;
+  }
+
+  &__message {
+    font-size: 1.2rem;
+    margin-bottom: 2rem;
+  }
+
+  &__list {
+    display: grid;
+    grid-template-columns: 300px 300px;
+    justify-content: center;
+    gap: 25px;
+    animation: appear 0.7s ;
+
+    @media (max-width: 700px) {
+      grid-template-columns: 250px;
+    }
+  }
+
+  &__item {
+    display: flex;
+    flex-direction: column;
+    animation: appear 0.7s ;
+    border: 1px solid black;
+    border-radius: 5px;
+    background: #dedddd;
+
+    box-shadow: 4px 3px 7px 0px rgba(0,0,0,0.75);
+    padding: 20px;
+    cursor: pointer;
+  }
+
+  &__text {
+    padding: 5px 0;
+    text-align: left;
+
+    &:first-child {
+      font-weight: bold;
+      line-height: 1.5rem;
+    }
+  }
+
+  &__link {
+    display: flex;
+    text-decoration: none;
+    flex-direction: column;
+    transition: 0.15s ease-in;
+
+    &:hover {
+      transform: scale(1.1) translateX(10px);
+    }
+  }
+
+  &__gitlink {
+    transition: 0.2s ease-in;
+
+    &:hover {
+      font-weight: bold;
+      color: red;
+    }
+  }
+}
+
+.loader {
+  border: 4px solid red;
+  border-top: 4px solid blue;
+  border-radius: 50%;
+  width: 120px;
+  height: 120px;
+  animation: spin 0.5s linear infinite;
+  margin: 100px auto;
+}
+
+@keyframes spin {
+
+  0% {
+    transform: rotate(0deg);
+  }
+  100% {
+    transform: rotate(360deg);
+  }
+}
+
+@keyframes appear {
+
+  0% {
+    opacity: 0;
+    transform: translateX(-150px);
+  }
+  100% {
+    opacity: 1;
+    transform: translateX(0px);
+  }
+
+}
