+++ src/components/ResultUser/index.ts
@@ -1,0 +1,3 @@
+import ResultUser from './ResultUser';
+
+export default ResultUser;
